fixes - remove receipts table for revert - committee storage in block refactoring
