+++ app/codes/minermanager.py
@@ -6,18 +6,20 @@ import logging
 # from app.codes.scoremanager import get_scores_for_wallets
 
 from .p2p.peers import add_peer
-from .clock.global_time import get_corrected_time_ms, get_time_difference
+
 from ..nvalues import SENTINEL_NODE_WALLET
 from .utils import get_last_block_hash
 # from .p2p.outgoing import propogate_transaction_to_peers
 from .p2p.utils import get_my_address
 from ..constants import COMMITTEE_SIZE, IS_TEST, MINIMUM_ACCEPTANCE_VOTES, NEWRL_DB, TIME_MINER_BROADCAST_INTERVAL_SECONDS
+from .clock.global_time import get_time_difference
 from .auth.auth import get_wallet
 from .signmanager import sign_transaction
 from ..ntypes import TRANSACTION_MINER_ADDITION
 from .utils import get_time_ms
 from .transactionmanager import Transactionmanager
 from .validator import validate
+from .committeemanager import get_eligible_miners, get_miner_for_current_block, get_committee_for_current_block
 
 
 logging.basicConfig(level=logging.INFO)
@@ -77,77 +79,6 @@ def get_my_miner_status():
 def broadcast_miner_update():
     transaction = miner_addition_transaction()
     validate(transaction, propagate=True)
-
-
-def get_eligible_miners():
-    # last_block = get_last_block_hash()
-    # last_block_epoch = 0
-    # try:
-    #     # Need try catch to support older block timestamps
-    #     last_block_epoch = int(last_block['timestamp'])
-    # except:
-    #     pass
-    # if last_block:
-    #     cutfoff_epoch = last_block_epoch - TIME_MINER_BROADCAST_INTERVAL
-    # else:
-    #     cutfoff_epoch = 0
-    # last_block_epoch = int(last_block['timestamp'])
-    cutfoff_epoch = get_corrected_time_ms() - TIME_MINER_BROADCAST_INTERVAL_SECONDS * 2 * 1000
-
-    con = sqlite3.connect(NEWRL_DB)
-    con.row_factory = sqlite3.Row
-    cur = con.cursor()
-    miner_cursor = cur.execute(
-        '''SELECT wallet_address, network_address, last_broadcast_timestamp 
-        FROM miners 
-        WHERE last_broadcast_timestamp > ?
-        ORDER BY wallet_address ASC''', (cutfoff_epoch, )).fetchall()
-    miners = [dict(m) for m in miner_cursor]
-    con.close()
-    return miners
-
-
-def get_miner_for_current_block(last_block=None):
-    if last_block is None:
-        last_block = get_last_block_hash()
-
-    if not last_block:
-        return {'wallet_address': SENTINEL_NODE_WALLET}
-
-    random.seed(last_block['index'])
-
-    committee_list = get_committee_for_current_block()
-
-    if len(committee_list) < MINIMUM_ACCEPTANCE_VOTES:
-        logger.info('Inadequate committee. Sentinel node is the miner.')
-        return {'wallet_address': SENTINEL_NODE_WALLET}
-
-    return random.choice(committee_list)
-
-    # return committee_list[0]
-
-
-def get_committee_for_current_block(last_block=None):
-    if last_block is None:
-        last_block = get_last_block_hash()
-
-    if not last_block:
-        return [{'wallet_address': SENTINEL_NODE_WALLET}]
-
-    random.seed(last_block['index'])
-
-    miners = get_eligible_miners()
-
-    if len(miners) == 0:
-        logger.info("No committee for current block. Using sentinel node.")
-        return [{'wallet_address': SENTINEL_NODE_WALLET}]
-
-    committee_size = min(COMMITTEE_SIZE, len(miners))
-    committee = random.sample(miners, k=committee_size)
-    # miner_wallets = list(map(lambda m: m['wallet_address'], miners))
-    # weights = get_scores_for_wallets(miner_wallets)
-    # committee = weighted_random_choices(miners, weights, committee_size)
-    return committee
 
 
 def get_committee_wallet_addresses():
